separate out report dataclasses for tests to pass
